Allow hiding model attributes in redirects
Context
-------

Spring Web MVC allows performing redirections by returning a String prefixed by `redirect://` in the controller.

This String is intercepted by `ThymeleafViewResolver` which maps it to an instance of `RedirectView`.

Problem
-------

The default behaviour of `RedirectView` is to include all model attributes as query parameters.

This behaviour may not be desirable and thymeleaf-spring does not provide a way to easily change it.

Suggested fix
-------------

Expose a method called `setRedirectExposeModelAttributes` in `ThymeleafViewResolver` to allow overriding this behaviour.